refactor: move deployment form models and validators to dedicated API model file
